refactor(native): clear the strict warning set
Fix all 42 production and 1 host-test warnings under
-Wall -Wextra -Wconversion -Wsign-conversion with explicit casts and
size_t boundaries. The generated code is unchanged: native 625d5300...
stays byte-identical, all 8 attack-critical functions match strictly
and full_cmp reports 642/642 shape-identical functions.

Enable the strict set in the Makefile and the CLion CMake project. Host
test binaries now depend on the production headers, which exposed and
fixed two stale tests that stopped compiling after the PiRace class
migration.

+++ src/core/kernelsnitch/futex_hash.h
@@ -240,7 +240,7 @@ futex_hash_context_bucket(const FutexHashContext *context, size_t addr,
 {
     futex_key_t key = {};
     key.private_key.mm = (void *)mm;
-    key.private_key.address = addr & ~0xfff;
+    key.private_key.address = addr & ~(size_t) 0xfff;
     key.private_key.offset = addr & 0xfff;
     return futex_hash_context_key(context, &key);
 }
