+++ src/core/runtime_struct_offsets.h
@@ -8,9 +8,9 @@ extern TargetProfile &g_target_profile;
 #define _RSO(field, fallback)                                                \
   target_profile_u32(&g_target_profile,                                     \
                      target_profile_values(&g_target_profile)               \
-                         ? target_profile_values(&g_target_profile)->field   \
+                         ? (uint32_t) target_profile_values(&g_target_profile)->field \
                          : 0,                                                \
-                     (fallback))
+                     (uint32_t) (fallback))
 #define _RSO_64(field, fallback) ((uint64_t)_RSO(field, fallback))
 #define _RSO_IMAGE(field, fallback) \
   (KIMAGE_TEXT_BASE + _RSO_64(field, fallback))
